+++ lib/ash/data_layer/data_layer.ex
@@ -15,7 +15,7 @@ defmodule Ash.DataLayer do
 
   @type feature() ::
           :transact
-          | :multitenant
+          | :multitenancy
           | {:lateral_join, list(Ash.Resource.t())}
           | {:join, Ash.Resource.t()}
           | {:aggregate, Ash.Query.Aggregate.kind()}