+++ lib/ash/query/ref.ex
@@ -1,6 +1,6 @@
 defmodule Ash.Query.Ref do
   @moduledoc "Represents a relation/attribute reference"
-  defstruct [:attribute, :relationship_path, :resource]
+  defstruct [:attribute, :relationship_path, :resource, :simple_equality?]
 
   defimpl Inspect do
     def inspect(ref, _opts) do